+++ src/Resources/contao/config/config.php
@@ -6,6 +6,7 @@
 $GLOBALS['TL_HOOKS']['getPageLayout']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'getPageLayout'];
 $GLOBALS['TL_HOOKS']['parseTemplate']['huh_amp'] = [\HeimrichHannot\AmpBundle\EventListener\ParseTemplateListener::class, 'onParseTemplate'];
 $GLOBALS['TL_HOOKS']['generatePage']['huh_amp'] = [\HeimrichHannot\AmpBundle\EventListener\GeneratePageListener::class, 'onGeneratePage'];
+$GLOBALS['TL_HOOKS']['parseFrontendTemplate']['huh_amp'] = [\HeimrichHannot\AmpBundle\EventListener\ParseFrontendTemplateListener::class, 'onParseFrontendTemplate'];
 $GLOBALS['TL_HOOKS']['modifyFrontendPage']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'modifyFrontendPage'];
 
 /**